requirements.txt
Flask==2.3.3
python-chess==1.999 